api/requirements.txt
fastapi[standard]
sqlmodel
python-jose
bcrypt
passlib
load_dotenv